Sequential Dynamical Invariant
The following video presents a demonstration of the modulation of robust yet flexible time interval relationships in a CPG network with varying electrical conductance for the Komendantov-Kononenko model. The video has been accelerated at specific moments in order to facilitate a more rapid illustration of the generation and modulation of the dynamical invariants as a function of the coupling conductance.